Add the `ProvableStorageCache::get_size_or_fetch` method (#2342)

+++ crates/module-system/sov-modules-api/src/state/tests.rs
@@ -25,7 +25,7 @@ fn test_charge_gas_to_set() {
     let gas_price = <<S as Spec>::Gas as Gas>::Price::from([1; 2]);
     let value = SlotValue::from("value");
     let gas_set_cost = <S as Spec>::Gas::from(config_value!("GAS_TO_CHARGE_PER_BYTE_FOR_WRITE"))
-        .checked_scalar_product(value.size() as u64)
+        .checked_scalar_product(value.size())
         .unwrap();
     let remaining_funds = gas_set_cost.value(&gas_price);
 
@@ -75,7 +75,7 @@ fn test_charge_gas_set_then_retrieve() {
 
     let value = SlotValue::from("value");
     let gas_set_cost = <S as Spec>::Gas::from(config_value!("GAS_TO_CHARGE_PER_BYTE_FOR_WRITE"))
-        .checked_scalar_product(value.size() as u64)
+        .checked_scalar_product(value.size())
         .unwrap();
 
     let remaining_funds = gas_access_cost.value(&gas_price) + gas_set_cost.value(&gas_price);
